backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.9.0
pydantic-settings==2.1.0
email-validator==2.2.0
faster-whisper
python-multipart
ffmpeg-python==0.2.0
numpy==1.24.3
pydub==0.25.1
moviepy
librosa
soundfile